scatclient/scatclient/tests/requirements.txt
pytest-html~=4.1.1
pytest-order
pytest-dependency
aiounittest